added login and register feature

+++ backend/controllers/imageController.js
@@ -5,7 +5,9 @@ const asyncHandler = require("express-async-handler");
 
 // GET Images
 const getImages = asyncHandler(async (req, res) => {
-    const allPhotos = await Images.find().sort({ createdAt: "descending" });
+    const allPhotos = await Images.find({ user: req.user.id }).sort({
+        createdAt: "descending",
+    });
     res.send(allPhotos);
 });
 
@@ -13,8 +15,9 @@ const getImages = asyncHandler(async (req, res) => {
 const createImage = asyncHandler(async (req, res) => {
     const photo = req.file.filename;
     const fileName = req.body.fileName;
+    const user = req.user.id;
 
-    const data = await Images.create({ photo, fileName });
+    const data = await Images.create({ user, photo, fileName });
     console.log("Uploaded Successfully...");
     console.log(data);
     res.send(data);
@@ -24,12 +27,32 @@ const createImage = asyncHandler(async (req, res) => {
 const deleteImage = asyncHandler(async (req, res) => {
     const id = req.params.id;
 
+    const image = await Images.findById(req.params.id)
+
+    // Check for user
+    if (!req.user) {
+        res.status(401);
+        throw new Error("User not found");
+    }
+    
+    // Make sure the logged in user matches the goal user
+    if (image.user.toString() !== req.user.id) {
+        res.status(401);
+        throw new Error("User not authorized");
+    }
+
     const data = await Images.findByIdAndRemove(id);
     if (!data) {
         return res.status(404).send("Image not found");
     }
 
-    const filePath = path.join(__dirname, "..", "public", "uploads", data.photo);
+    const filePath = path.join(
+        __dirname,
+        "..",
+        "public",
+        "uploads",
+        data.photo
+    );
 
     if (fs.existsSync(filePath)) {
         fs.unlink(filePath, (err) => {
